pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "aefimenko"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "uuid6 (>=2025.0.1,<2026.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "autoflake (>=2.3.1,<3.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)",
    "yamllint (>=1.37.1,<2.0.0)"
]


[tool.black]
target-version = ["py313"]
line-length = 120
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.tox
    | \.venv
    | venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''